// management/frontend/src/api/schemas/collectionTaskPublic.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Ego Collection Management
 * OpenAPI spec version: 0.1.0
 */

export interface CollectionTaskPublic {
  id?: string;
  created_at?: string | null;
  updated_at?: string | null;
  deleted_at?: string | null;
  creator_id?: string | null;
  updater_id?: string | null;
  deleter_id?: string | null;
  /** @maxLength 64 */
  task_no?: string;
  /**
     * @minLength 1
     * @maxLength 128
     */
  project_name: string;
  /**
     * @minLength 1
     * @maxLength 128
     */
  name: string;
  /** @maxLength 128 */
  subtask_name?: string;
  /** @maxLength 64 */
  scene_type?: string;
  kit_id?: string | null;
  sop_id: string;
  /** @maxLength 128 */
  assigned_username?: string;
  /** @maxLength 128 */
  device_serial?: string;
  /** @maxLength 256 */
  location?: string;
  /** @maxLength 1000 */
  target_objects?: string;
  /** @minimum 0 */
  object_count?: number;
  /** @minimum 0 */
  duration_minutes?: number;
  /** @minimum 0 */
  target_count?: number;
  /** @minimum 0 */
  completed_count?: number;
  published_at?: string | null;
  /** @maxLength 32 */
  status?: string;
  workspace_id: string;
  kit_name?: string;
  sop_name?: string;
  sop_content?: string;
  creator_name?: string | null;
  updater_name?: string | null;
}